Drew the MainMenu GUI for testing.
