+++ ALES Project/src/utility/GUIs.java
@@ -3,11 +3,8 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package sim;
+package utility;
 
-import gui.GUI;
-import java.util.HashMap;
-import java.util.Map;
 import util.Color4;
 import util.Vec2;
 
@@ -17,15 +14,8 @@ import util.Vec2;
  */
 public class GUIs {
 
-    private static final Map<String, GUI> GUIS = new HashMap();
-    private static final Color4[] COLORS = {Color4.BLUE.withG(1.0), Color4.RED, Color4.WHITE};
-    
+    public static final Color4[] COLORS = {Color4.ORANGE, Color4.BLUE, Color4.WHITE};
     public static final Vec2 BUTTON_SIZE = new Vec2(200, 50);
-
-    static {
-
-        GUIS.put("mainMenu", new MainMenu("mainMenu"));
-    }
 
     public static Color4 getColor(int rank) {
